refactor(reservations): finish command flow cleanup
