+++ apps/server/src/domain/reservations/index.ts
@@ -3,14 +3,13 @@ export * from "./models";
 export * from "./repository/reservation-analytics.repository";
 export * from "./repository/reservation-command.repository";
 export * from "./repository/reservation-query.repository";
-export * from "./services/cancel-reservation.service";
+export * from "./services/commands/cancel-reservation.service";
 export * from "./services/commands/confirm-reservation.service";
 export * from "./services/commands/reserve-bike.service";
 export * from "./services/fixed-slot-template.service";
 export * from "./services/fixed-slot/fixed-slot.helpers";
 export * from "./services/fixed-slot/fixed-slot.service";
 export type * from "./services/fixed-slot/fixed-slot.types";
-export * from "./services/reservation-command.service";
 export * from "./services/reservation-query.service";
 export * from "./services/reservation-stats.service";
 export * from "./types";
